requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning & Computer Vision
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
facenet-pytorch==2.5.3
pillow==10.0.1
numpy==1.24.3

# Data Processing
pandas==2.1.1
scikit-learn==1.3.1

# HTTP Client for testing
httpx==0.25.1
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
